ai_service/requirements.txt
ultralytics
opencv-python
numpy
fastapi
uvicorn
pillow
python-multipart
python-dotenv
httpx
